ulitmos ajustes para o deploy
ulitmos ajustes para o deploy

+++ backend/models/servico.py
@@ -5,5 +5,5 @@ class Servico(db.Model):
 
     id = db.Column(db.Integer,primary_key = True)
     nome_servico = db.Column(db.String(150),nullable = False)
-    valor = db.Column(db.Float(100),nullable = False)
+    valor = db.Column(db.Float, nullable = False)
     id_barbeiro = db.Column(db.Integer, db.ForeignKey('barbeiros.id'),nullable = False)
